chore: refactor stage sidebar
